gateway: activate BOOTCOUNT support
If no softreset save registers are found on the hardware "bootcount"
is stored in the environment. To prevent a saveenv on all reboots, the
environment variable "upgrade_available" is used. If
"upgrade_available" is 0, "bootcount" is always 0. If
"upgrade_available" is 1, "bootcount" is incremented in the
environment.

So the Userspace Application must set the "upgrade_available"
and "bootcount" variables to 0, if the system booted successfully.